+++ client/src/shared/hooks/useEnterKeyHandler.ts
@@ -1,0 +1,15 @@
+import { useEffect } from 'react';
+
+export const useEnterKeyHandler = (callback?: () => void) => {
+  useEffect(() => {
+    const handleKeyDown = (e: KeyboardEvent) => {
+      if (e.key === 'Enter' && callback) {
+        e.preventDefault();
+        callback();
+      }
+    };
+
+    document.addEventListener('keydown', handleKeyDown);
+    return () => document.removeEventListener('keydown', handleKeyDown);
+  }, [callback]);
+};
