+++ client/src/widgets/signup/SignupStep3.tsx
@@ -1,6 +1,6 @@
 import { SignupFormData } from '@/features/auth/types/signup';
 import { Button } from '@/shared/ui/button/Button';
-import { useEffect } from 'react';
+import { useEnterKeyHandler } from '@/shared/hooks/useEnterKeyHandler';
 import * as S from './SignupStep3.styles';
 
 interface SignupStep3Props {
@@ -10,17 +10,7 @@ interface SignupStep3Props {
 }
 
 export const SignupStep3 = ({ signupData, handleClick, onEnter }: SignupStep3Props) => {
-  useEffect(() => {
-    const handleKeyDown = (e: KeyboardEvent) => {
-      if (e.key === 'Enter' && onEnter) {
-        e.preventDefault();
-        onEnter();
-      }
-    };
-
-    document.addEventListener('keydown', handleKeyDown);
-    return () => document.removeEventListener('keydown', handleKeyDown);
-  }, [onEnter]);
+  useEnterKeyHandler(onEnter);
 
   return (
     <S.StepContainer>
